Allow output logging from fmt or log packages when executable is running.
As noted in issue #50, output logging only shows up when `air` is closed.  This is a bit annoying since you have to close `air` (and most likely restart it) to see any log messages.  This change fixes the output to occur while `air` is running and thus mirrors the functionality in  `fresh`.  I am unsure of any unintended side-effects of this change, but I have not noticed any thus far.

It looks like this bug/unintended functionality/whatever-you-want-to-call-it is caused by the wrapping of the `io.Copy()` calls for stdout and stderr into one goroutine and using with the `cmd.Process.Wait()` call.  The `Wait()` call hangs/blocks endlessly until the binary (i.e. `air`) is closed/stopped thus not allowing the copying of the stdout or stderr to the terminal you ran `air` in.  However, just removing the `Wait()` call did not allow logging to work; the `Copy()` calls had to be split into separate goroutines as well.  Thus, at the end of the day, the code to fix this is simply a copying of the code from `fresh`.

References:
 - `fresh` lines https://github.com/gravityblast/fresh/blob/master/runner/runner.go#L28-L29.
 - `air` lines https://github.com/cosmtrek/air/blob/6f18a44eb4aa1eac4956755d7efc3605fc42f2ac/runner/engine.go#L434-L437